cria_sigla movido para utils/
